import electron dialog refactoring
