+++ src/ts/interface/popup.ts
@@ -24,6 +24,5 @@ export interface PopupSettings extends Popup {
 	setConfirmPin: (v: () => void) => void;
 	setPinConfirmed: (v: boolean) => void;
 	onConfirmPin: () => void;
-	onImport: (type: I.ImportType, param: any, callBack?: (message: any) => void) => void;
 	onExport: (format: I.ExportType, param: any) => void;
 };
