+++ src/ts/lib/action.ts
@@ -2,7 +2,6 @@ import { I, C, focus, analytics, Renderer, Preview, UtilCommon, Storage, UtilDat
 import { commonStore, authStore, blockStore, detailStore, dbStore, popupStore } from 'Store';
 import Constant from 'json/constant.json';
 
-
 class Action {
 
 	pageClose (rootId: string, close: boolean) {
@@ -396,6 +395,40 @@ class Action {
 			if (callBack) {
 				callBack();
 			};
+		});
+	};
+
+	import (type: I.ImportType, extensions: string[], options?: any, callBack?: (message: any) => void) {
+		const fileOptions: any = { 
+			properties: [ 'openFile' ],
+			filters: [ 
+				{ name: 'Filtered extensions', extensions },
+			],
+		};
+
+		if (UtilCommon.isPlatformMac()) {
+			fileOptions.properties.push('openDirectory');
+		};
+
+		analytics.event('ClickImport', { type });
+
+		window.Electron.showOpenDialog(fileOptions).then((result: any) => {
+			const paths = result.filePaths;
+			if ((paths == undefined) || !paths.length) {
+				return;
+			};
+
+			analytics.event('ClickImportFile', { type });
+
+			C.ObjectImport(Object.assign(options || {}, { paths }), [], true, type, I.ImportMode.IgnoreErrors, false, false, (message: any) => {
+				if (!message.error.code) {
+					analytics.event('Import', { middleTime: message.middleTime, type });
+				};
+
+				if (callBack) {	
+					callBack(message);
+				};
+			});
 		});
 	};
